requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
alembic==1.13.1
anyio==4.3.0
apache-airflow==2.8.3
apache-airflow-providers-common-io==1.3.0
apache-airflow-providers-common-sql==1.11.1
apache-airflow-providers-ftp==3.7.0
apache-airflow-providers-http==4.10.0
apache-airflow-providers-imap==3.5.0
apache-airflow-providers-smtp==1.6.1
apache-airflow-providers-sqlite==3.7.1
apispec==6.5.0
argcomplete==3.2.3
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
Babel==2.14.0
bcc==0.18.0
beautifulsoup4==4.10.0
blinker==1.4
Brlapi==0.8.3
bs4==0.0.2
cachelib==0.9.0
certifi==2020.6.20
chardet==4.0.0
charset-normalizer==3.3.2
chrome-gnome-shell==0.0.0
click==8.0.3
clickclick==20.10.2
colorama==0.4.4
colorlog==4.8.0
command-not-found==0.3
ConfigUpdater==3.2
connexion==2.14.2
cron-descriptor==1.4.3
croniter==2.0.2
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
Deprecated==1.2.14
dill==0.3.8
distlib==0.3.4
distro==1.7.0
dnspython==2.6.1
docutils==0.20.1
email-validator==1.3.1
exceptiongroup==1.2.0
filelock==3.6.0
Flask==2.2.5
Flask-AppBuilder==4.3.11
Flask-Babel==2.0.0
Flask-Caching==2.1.0
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.5.1
Flask-Login==0.6.3
Flask-Session==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.2.1
frozenlist==1.4.1
fsspec==2024.2.0
google-re2==1.1
googleapis-common-protos==1.63.0
greenlet==3.0.3
grpcio==1.62.1
gunicorn==21.2.0
h11==0.14.0
hidpidaemon==18.4.6
html5lib==1.1
httpcore==1.0.4
httplib2==0.20.2
httpx==0.27.0
idna==3.3
importlib-metadata==6.11.0
importlib_resources==6.1.3
inflection==0.5.1
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy-object-proxy==1.10.0
limits==3.10.0
linkify-it-py==2.0.3
lockfile==0.12.2
louis==3.20.0
lxml==4.8.0
macaroonbakery==1.3.1
Mako==1.3.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.1
marshmallow-oneofschema==3.1.1
marshmallow-sqlalchemy==0.26.1
mdit-py-plugins==0.4.0
mdurl==0.1.2
more-itertools==10.2.0
multidict==6.0.5
netaddr==0.8.0
netifaces==0.11.0
numpy==1.26.4
oauthlib==3.2.0
opentelemetry-api==1.23.0
opentelemetry-exporter-otlp==1.23.0
opentelemetry-exporter-otlp-proto-common==1.23.0
opentelemetry-exporter-otlp-proto-grpc==1.23.0
opentelemetry-exporter-otlp-proto-http==1.23.0
opentelemetry-proto==1.23.0
opentelemetry-sdk==1.23.0
opentelemetry-semantic-conventions==0.44b0
ordered-set==4.1.0
packaging==23.2
pandas==2.2.1
pathspec==0.12.1
pendulum==3.0.0
platformdirs==2.5.1
pluggy==1.4.0
pop-transition==1.1.2
prison==0.2.1
protobuf==4.25.3
psutil==5.9.8
pyarrow==15.0.1
pycairo==1.20.1
pycups==2.0.1
pydbus==0.6.0
Pygments==2.17.2
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
python-apt==2.4.0+ubuntu3
python-daemon==3.0.1
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-gnupg==0.4.8
python-nvd3==0.15.0
python-slugify==8.0.4
python-xlib==0.29
pytz==2024.1
pyxdg==0.27
PyYAML==5.4.1
referencing==0.33.0
repolib==2.2.1
repoman==1.4.0
requests==2.31.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rich==13.7.1
rich-argparse==1.4.0
rpds-py==0.18.0
SecretStorage==3.3.1
sessioninstaller==0.0.0
setproctitle==1.3.3
six==1.16.0
sniffio==1.3.1
soupsieve==2.3.1
SQLAlchemy==1.4.52
SQLAlchemy-JSONField==1.0.2
SQLAlchemy-Utils==0.41.1
sqlparse==0.4.4
systemd-python==234
tabulate==0.9.0
tenacity==8.2.3
termcolor==2.4.0
text-unidecode==1.3
time-machine==2.14.0
typing_extensions==4.10.0
tzdata==2024.1
ubuntu-drivers-common==0.0.0
uc-micro-py==1.0.3
ufw==0.36.1
unicodecsv==0.14.1
universal-pathlib==0.1.4
urllib3==1.26.5
virtualenv==20.13.0+ds
wadllib==1.3.6
webencodings==0.5.1
Werkzeug==2.2.3
wrapt==1.16.0
WTForms==3.1.2
xdg==5
xkit==0.0.0
yarl==1.9.4
zipp==1.0.0
